Make size arg of r_file_slurp() size_t

+++ libr/asm/asm.c
@@ -155,13 +155,13 @@ static inline int r_asm_pseudo_fill(RAsmOp *op, char *input) {
 }
 
 static inline int r_asm_pseudo_incbin(RAsmOp *op, char *input) {
-	int bytes_read = 0;
+	size_t bytes_read = 0;
 	r_str_replace_char (input, ',', ' ');
 	// int len = r_str_word_count (input);
 	r_str_word_set0 (input);
 	//const char *filename = r_str_word_get0 (input, 0);
-	int skip = (int)r_num_math (NULL, r_str_word_get0 (input, 1));
-	int count = (int)r_num_math (NULL,r_str_word_get0 (input, 2));
+	size_t skip = (size_t)r_num_math (NULL, r_str_word_get0 (input, 1));
+	size_t count = (size_t)r_num_math (NULL,r_str_word_get0 (input, 2));
 	char *content = r_file_slurp (input, &bytes_read);
 	if (!content) {
 		eprintf ("Could not open '%s'.\n", input);
